auth slice fixed
