+++ src/pages/Profile/Profile.jsx
@@ -1,23 +1,18 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 
-// import { SharedLayout } from 'components/SharedLayout/SaredLayout';
-import { HiCamera, HiTrash } from 'react-icons/hi2';
+import { HiTrash } from 'react-icons/hi2';
 import { HiOutlineLogout } from 'react-icons/hi';
 import { BsPlusCircleFill } from 'react-icons/bs';
 import {
   Section,
-  H2,
+  UserPartTitle,
   UserInfo,
-  AvatarContainer,
-  Avatar,
-  ChangeAvatarForm,
-  AvatarButton,
-  EditAvatarText,
-  UserDataContainer,
+  UserData,
   LogOutContainer,
   LogOutButton,
   LogOutText,
   PetsHeader,
+  PetsPartTitle,
   AddPetContainer,
   AddPetButton,
   AddPetText,
@@ -28,9 +23,10 @@ import {
   Span,
   P,
 } from './Profile.styled';
-import { UserUpdateForm } from 'components/UserUpdateForm/UserUdateForm';
+import { Modal } from 'components/Modal/Modal';
+import { Avatar } from 'components/Profile/Avatar/Avatar';
+import { UserUpdateForm } from 'components/Profile/UserUpdateForm/UserUdateForm';
 
-import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { selectToken } from 'redux/Auth/selectors';
 import {
@@ -41,7 +37,6 @@ import {
 } from 'services/api/user';
 import { logout } from 'redux/Auth/operations';
 import { useNavigate } from 'react-router';
-import { Modal } from 'components/Modal/Modal';
 
 const convertDate = date => {
   const dateOptions = { day: 'numeric', month: 'numeric', year: 'numeric' };
@@ -85,17 +80,9 @@ export const Profile = () => {
     setUserData(prevState => ({ ...prevState, ...data }));
   };
 
-  const onAvatarInputChange = async event => {
-    const avatar = event.currentTarget.files[0];
-
-    const reader = new FileReader();
-    reader.readAsDataURL(avatar);
-    reader.onload = event =>
-      setUserData(prevState => ({
-        ...prevState,
-        avatarURL: event.target.result,
-      }));
+  const changeAvatar = (avatar, avatarUrl) => {
     uploadAvatar(avatar, token);
+    setUserData(prevState => ({ ...prevState, avatarURL: avatarUrl }));
   };
 
   const logoutUser = event => {
@@ -118,39 +105,18 @@ export const Profile = () => {
   return (
     <>
       <Section>
-        {/* ------------------------ USER ------------------------ */}
-
+        {/* ------------------------ USER PART ------------------------ */}
         <section>
-          <H2>My information:</H2>
+          <UserPartTitle>My information:</UserPartTitle>
           <UserInfo>
-            {/* -------------------- AVATAR ---------------------- */}
+            {userData.avatarURL && (
+              <Avatar
+                avatarURL={userData.avatarURL}
+                changeAvatar={changeAvatar}
+              />
+            )}
 
-            <AvatarContainer>
-              <Avatar>
-                <img src={userData.avatarURL} alt="avatar" />
-              </Avatar>
-              <ChangeAvatarForm>
-                <AvatarButton htmlFor="upload-avatar">
-                  <HiCamera size={20} color={'#F59256'} />
-                </AvatarButton>
-                <input
-                  type="file"
-                  id="upload-avatar"
-                  onChange={onAvatarInputChange}
-                  style={{
-                    opacity: 0,
-                    position: 'absolute',
-                    zIndex: -1,
-                    pointerEvents: 'none',
-                  }}
-                />
-                <EditAvatarText>Edit photo</EditAvatarText>
-              </ChangeAvatarForm>
-            </AvatarContainer>
-
-            {/* ------------------- USER INFO --------------------- */}
-
-            <UserDataContainer>
+            <UserData>
               {Object.keys(userData).length !== 0 && (
                 <UserUpdateForm
                   data={userData}
@@ -158,9 +124,7 @@ export const Profile = () => {
                   token={token}
                 />
               )}
-            </UserDataContainer>
-
-            {/* ------------------ LOG OUT BTN --------------------- */}
+            </UserData>
 
             <LogOutContainer>
               <LogOutButton type="button" onClick={logoutUser}>
@@ -171,11 +135,13 @@ export const Profile = () => {
           </UserInfo>
         </section>
 
-        {/* ------------------------ PETS ------------------------ */}
+        {/* ------------------------ PETS PART ------------------------ */}
 
         <section>
           <PetsHeader>
-            <H2 style={{ marginBottom: '0px' }}>My pets:</H2>
+            <PetsPartTitle style={{ marginBottom: '0px' }}>
+              My pets:
+            </PetsPartTitle>
             <AddPetContainer>
               <AddPetText>Add Pet</AddPetText>
               <AddPetButton type="button" onClick={toggleModal}>
@@ -183,8 +149,6 @@ export const Profile = () => {
               </AddPetButton>
             </AddPetContainer>
           </PetsHeader>
-
-          {/* -------------------- PET AVATAR --------------------- */}
 
           <ul>
             {userPets.map((pet, index) => (
@@ -208,8 +172,6 @@ export const Profile = () => {
                   <P>
                     <Span>Comments: </Span> {pet.comments}
                   </P>
-
-                  {/* -------------DELETE PET BTN */}
 
                   <DeletePetButton
                     type="button"
